Revision of the architecture
Reduce coupling  in the InternetBankController and BankAccountService:
* Create ManagementInternetBankController.
  Methods deleteAccount() and createAccount() from
  InternetBankController have been transferred to it;
* Create OperationsController.
  Method getOperationList() from InternetBankController have been
  transferred to it.
* Create service BankAccountManager.
  Methods deleteAccount() and createAccount() from BankAccountService
  have been transferred to it.
